requirements.txt
fastapi
uvicorn
mangum
anthropic
python-dotenv
